Renamed the paint sidebar to paint-toolbar since it now changes orientation. #22
