+++ client/utils/paint-toolbar.js
